+++ flycheck.el
@@ -978,8 +978,11 @@ to a number and return it.  Otherwise return nil."
 
 (defun flycheck-same-files-p (file-a file-b)
   "Determine whether FILE-A and FILE-B refer to the same file."
-  (string= (directory-file-name (expand-file-name file-a))
-           (directory-file-name (expand-file-name file-b))))
+  ;; We must resolve symbolic links here, since some syntax checker always
+  ;; output canonical file names with all symbolic links resolved.  See
+  ;; https://github.com/flycheck/flycheck/issues/561
+  (string= (directory-file-name (file-truename file-a))
+           (directory-file-name (file-truename file-b))))
 
 (defvar-local flycheck-temporaries nil
   "Temporary files and directories created by Flycheck.")